Move Source into platform module
