+++ crates/volta-core/src/lib.rs
@@ -21,7 +21,6 @@ pub mod project;
 pub mod session;
 pub mod shell;
 pub mod shim;
-pub mod source;
 pub mod style;
 pub mod tool;
 pub mod toolchain;
