--- crates/volta-core/src/source.rs
@@ -1,5 +1,0 @@
-#[derive(Debug)]
-pub enum Source {
-    User,
-    Project,
-}
