+++ crates/volta-core/src/tool/binary.rs
@@ -3,8 +3,8 @@ use std::iter::once;
 
 use super::ToolCommand;
 use crate::error::ErrorDetails;
+use crate::platform::Source;
 use crate::session::{ActivityKind, Session};
-use crate::source::Source;
 
 use log::debug;
 use volta_fail::{throw, Fallible};
